# docs/swagger.yaml
definitions:
...
  comments.Comment:
    properties:
      author:
        type: string
      authorId:
        type: integer
      created:
        type: string
      id:
        type: integer
      postId:
        type: integer
      text:
        type: string
    type: object
...
  posts.Post:
    properties:
      author:
        type: string
      authorId:
        type: integer
      comments:
        type: integer
      created:
        type: string
      id:
        type: integer
      likes:
        type: integer
      tags:
        items:
          $ref: '#/definitions/tags.Tag'
        type: array
      text:
        type: string
      title:
        type: string
    type: object
  posts.Posts:
    properties:
      nposts:
        type: integer
      posts:
        items:
          $ref: '#/definitions/posts.Post'
        type: array
    type: object
  tags.Tag:
    properties:
      name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/comments/post/{id}:
    get:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/comments.Comment'
            type: array
      summary: Get comments for the post
      tags:
      - comments
...
  /api/posts/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/posts.Post'
      - description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: Add a new post
      tags:
      - posts
...
  /api/tags/post/{name}:
    get:
      parameters:
      - description: Tag Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.Posts'
      summary: Get posts associated with the tag
      tags:
      - tags
swagger: "2.0"